relocate some test data
