+++ tests/test_database_plugin.py
@@ -12,57 +12,13 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-import logging
 import unittest
 
 from sqlalchemy.exc import IntegrityError
 
-from plugins.providers.sql import ServerDataModel, SwitchInfo
-from plugins.providers.sql import SQL
-from tests import Config
 from tests import PigTests
-config = Config()
 
 class TestSQL(PigTests):
-
-    @classmethod
-    def setUpClass(cls):
-        cls.sql = SQL(config)
-        logging.basicConfig()
-        log = logging.getLogger(__name__)
-        log.setLevel(logging.DEBUG)
-        cls._add_entry()
-
-    @classmethod
-    def _add_entry(cls):
-        server_entry = ServerDataModel()
-        server_entry.boot_os = "Ubuntu"
-        server_entry.boot_os_version = "14.04.2"
-        server_entry.boot_profile = "Unknown"
-        server_entry.boot_status = "Kicking"
-        server_entry.bootstrapped = False
-        server_entry.dns_domain_name = "rpc.local"
-        server_entry.dns_server_primary = "8.8.8.8"
-        server_entry.hostname = "hogzilla"
-        server_entry.operational_status = "Provisioning"
-        server_entry.primary_gw = "10.12.1.1"
-        server_entry.primary_ip = "10.12.1.10"
-        server_entry.primary_mac = "00:11:22:33:44:55"
-        server_entry.primary_nm = "255.255.255.0"
-        server_entry.server_number = 555121
-        cls.sql.create_entry(server_entry)
-
-        switch_entry = SwitchInfo()
-        switch_entry.server_number = 555121
-        switch_entry.switch_name = "Switch 01"
-        switch_entry.switch_port = "1"
-        cls.sql.add_switch_entry(switch_entry)
-
-        switch_entry1 = SwitchInfo()
-        switch_entry1.server_number = 555121
-        switch_entry1.switch_name = "Switch 01"
-        switch_entry1.switch_port = "2"
-        cls.sql.add_switch_entry(switch_entry1)
 
     def test_duplicate_server_number_not_allowed(self):
         with self.assertRaises(IntegrityError):
